added separate unit test file (Team)
Separated unit tests into its own python file
